app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
binaryornot==0.4.4
briefcase==0.3.23
build==1.2.2.post1
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
cookiecutter==2.6.0
fastapi==0.115.12
flet==0.28.2
flet-cli==0.28.2
flet-desktop==0.28.2
flet-web==0.28.2
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.2.2
packaging==25.0
platformdirs==4.3.8
psutil==7.0.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pypng==0.20220715.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
rich==14.0.0
setuptools==80.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
text-unidecode==1.3
toml==0.10.2
tomli_w==1.2.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchdog==4.0.2
watchfiles==1.0.5
websockets==15.0.1
wheel==0.45.1
